UserService/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
aio-pika
python-dotenv
bcrypt
pyodbc